clean avail. imports
