+++ pytorch_lightning/utilities/apply_func.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import importlib
 from abc import ABC
 from collections.abc import Mapping, Sequence
 from copy import copy
@@ -20,8 +19,9 @@ from typing import Any, Callable, Union
 
 import torch
 
-TORCHTEXT_AVAILABLE = importlib.util.find_spec("torchtext") is not None
-if TORCHTEXT_AVAILABLE:
+from pytorch_lightning.utilities.imports import _TORCHTEXT_AVAILABLE
+
+if _TORCHTEXT_AVAILABLE:
     from torchtext.data import Batch
 else:
     Batch = type(None)
@@ -109,7 +109,7 @@ def move_data_to_device(batch: Any, device: torch.device):
 
     def batch_to(data):
         # try to move torchtext data first
-        if TORCHTEXT_AVAILABLE and isinstance(data, Batch):
+        if _TORCHTEXT_AVAILABLE and isinstance(data, Batch):
 
             # Shallow copy because each Batch has a reference to Dataset which contains all examples
             device_data = copy(data)
@@ -123,5 +123,5 @@ def move_data_to_device(batch: Any, device: torch.device):
         kwargs = dict(non_blocking=True) if isinstance(data, torch.Tensor) else {}
         return data.to(device, **kwargs)
 
-    dtype = (TransferableDataType, Batch) if TORCHTEXT_AVAILABLE else TransferableDataType
+    dtype = (TransferableDataType, Batch) if _TORCHTEXT_AVAILABLE else TransferableDataType
     return apply_to_collection(batch, dtype=dtype, function=batch_to)
